mythexternrecorder: rename commandlineparser.* to mythexternrecorder_…
