+++ mythtv/programs/mythexternrecorder/mythexternrecorder.cpp
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "commandlineparser.h"
+#include "mythexternrecorder_commandlineparser.h"
 #include "MythExternRecApp.h"
 #include "MythExternControl.h"
