+++ mythtv/programs/mythexternrecorder/mythexternrecorder_commandlineparser.h
